refactor: extract tier types and refactor their names
