+++ Applications/AdminUi/packages/admin_api_sdk/lib/src/endpoints/tiers_endpoint.dart
@@ -6,9 +6,9 @@ import 'endpoint.dart';
 class TiersEndpoint extends Endpoint {
   TiersEndpoint(super.dio);
 
-  Future<ApiResponse<List<TierOverview>>> getTiers() => get(
+  Future<ApiResponse<List<TierOverviewResponse>>> getTiers() => get(
         '/api/v1/Tiers',
-        transformer: (e) => (e as List).map(TierOverview.fromJson).toList(),
+        transformer: (e) => (e as List).map(TierOverviewResponse.fromJson).toList(),
       );
 
   Future<ApiResponse<Tier>> createTier({required String name}) => post(
