+++ Applications/AdminUi/packages/admin_api_types/lib/src/tiers/tier_details.dart
@@ -1,0 +1,21 @@
+import 'package:json_annotation/json_annotation.dart';
+
+import '../quotas/quotas.dart';
+
+part 'tier_details.g.dart';
+
+@JsonSerializable()
+class TierDetails {
+  final String id;
+  final String name;
+  final List<Quota> quotas;
+
+  TierDetails({
+    required this.id,
+    required this.name,
+    required this.quotas,
+  });
+
+  factory TierDetails.fromJson(dynamic json) => _$TierDetailsFromJson(json as Map<String, dynamic>);
+  Map<String, dynamic> toJson() => _$TierDetailsToJson(this);
+}
